Solve boundaries overflow issues
When calculating the new end, there was a chance than we get < than 0 or simply that we miss the word. e.g. the word we are searching is right before and the current middle word, then because of the -1 that was used for calculating the end would cause that we skip the word completly